# doc/api.yaml
openapi: 3.0.3
info:
  title: WasaPhotoAPI Specification
  description: |
    This OpenAPI document describes the WasaPhoto project REST API.

    WasaPhoto is a social media platform that focuses on photo sharing.

  version: 1.0.0
servers:
  - url: "http://localhost:3000"
    description: localhost server for development
security:
  - bearerAuthentication: []
...
paths:
...
  /follow/{userId}:
    put:
      tags: ["follow"]
      operationId: followUser
      summary: Follows a user.
      description: |-
        Follows the user specified with the UserId.
      parameters: 
        - $ref: "#/components/parameters/UserId" 
      responses:
        "204": { $ref: "#/components/responses/NoContent" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "500": { $ref: "#/components/responses/InternalServerError" }

    delete:
      tags: ["follow"]
      operationId: unfollowUser
      summary: Unfollows a user.
      description: |-
        Unfollows the user specified with the UserId.
      parameters: 
        - $ref: "#/components/parameters/UserId" 
      responses:
        "204": { $ref: "#/components/responses/NoContent" } #unfollowed the user
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "500": { $ref: "#/components/responses/InternalServerError" }
  /followers/{userId}:
    get:
      tags: ["follow"]
      operationId: getFollowers
      summary: Gets the list of followers.
      description: |-
        Gets the list of followers. The user that blocked the requester will not be shown.
      parameters: 
        - $ref: "#/components/parameters/UserId" 
      responses:
        "200": 
          description: Returns the list of the users that follows the logged user.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/UserList" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "500": { $ref: "#/components/responses/InternalServerError" }
  /following/{userId}:
    get:
      tags: ["follow"]
      operationId: getFollowing
      summary: Gets the list of following.
      description: |-
        Gets the list of following. The user that blocked the requester will not be shown.
      parameters: 
        - $ref: "#/components/parameters/UserId" 
      responses:
        "200": 
          description: Returns the list of the users that the logged user follows.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/UserList" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "500": { $ref: "#/components/responses/InternalServerError" }

#-----------------------------
components:
  securitySchemes:
    bearerAuthentication:
      description: |
        User token that uses the entityId of the logged user.
        Done accordingly to the project specification
      type: http
      scheme: bearer    
  schemas:
...
    UserReduced:
      description: |
        The Users entity with only the essential properties.
      type: object
      properties:
        userId:
          type: integer
          example: 1
        username:
          type: string
          example: "Username"
    UserList:
      type: array
      description: List of Users
      items: 
        $ref: "#/components/schemas/UserReduced"
...
  parameters:
...
    UserId:
      in: path
      name: userId
      description: The EntityId of the requested user.
      schema:
        type: integer
        example: 1
        maximum: 18446744073709551615 
        minimum: 0
      required: true
...
  responses:
...
    NoContent:
      description: The request ended successfully with no body
    Unauthorized:
      description: The access token is missing or it's expired
    BadRequest:
      description: The request was not compliant with the documentation (eg. missing fields, etc)
    InternalServerError:
      description: The server encountered an internal error. Further info in server logs.
    NotFound:
      description: The requested entity is not found.
...